requirements.txt
git+https://github.com/Kyrela/discore
python-i18n~=0.3.9
psutil~=5.9.5
requests~=2.31.0
masonite-orm~=2.19.2
pymysql~=1.1.0
